fixed uint32_t error? changed to uintptr_t, see https://stackoverflow.com/questions/21232460/disabling-cast-from-pointer-to-smaller-type-uint32-t-error-in-clang

+++ elements/local/igmp/IGMPClientSide.cc
@@ -172,7 +172,7 @@ WritablePacket * IGMPClientSide::make_mem_report_packet()
         igmp_grp->record_type = record.record_type;
 
         // add source adresses on top
-        uint32_t igmp_adr = (uint32_t) (igmp_grp + 1);
+        uintptr_t igmp_adr = (uintptr_t) (igmp_grp + 1);
         for (IPAddress adress:record.sources)
         {
             igmp_adr = adress.addr();
